# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Home Logbook API
  version: 0.11.0
servers:
- url: http://192.168.68.62:8787/api/v1
security:
- bearerAuth: []
paths:
...
  /unit-types:
    get:
      summary: List unit types and units
      responses:
        '200':
          description: Unit catalog
          content:
            application/json:
              schema:
                type: object
                required:
                - unitTypes
                properties:
                  unitTypes:
                    type: array
                    items:
                      $ref: '#/components/schemas/UnitType'
    post:
      summary: Create a unit type and its canonical base unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitTypeCreate'
      responses:
        '201':
          description: Created
        '409':
          description: Duplicate key
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /unit-types/{key}:
    parameters:
    - in: path
      name: key
      required: true
      schema:
        type: string
    get:
      summary: Get one unit type
      responses:
        '200':
          description: Unit type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitType'
        '404':
          description: Unknown unit type
    patch:
      summary: Rename or describe a unit type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type:
                  - string
                  - 'null'
      responses:
        '200':
          description: Updated
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
    delete:
      summary: Delete an entirely unused unit type
      responses:
        '204':
          description: Deleted
        '409':
          description: Referenced by an event type or event
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /unit-types/{key}/units:
    parameters:
    - in: path
      name: key
      required: true
      schema:
        type: string
    post:
      summary: Add a unit with conversion back to the base unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitCreate'
      responses:
        '201':
          description: Added
        '409':
          description: Duplicate unit key
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /unit-types/{key}/units/{unitKey}:
    parameters:
    - in: path
      name: key
      required: true
      schema:
        type: string
    - in: path
      name: unitKey
      required: true
      schema:
        type: string
    patch:
      summary: Update a unit
      description: Conversion factors are immutable after the unit is used.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitUpdate'
      responses:
        '200':
          description: Updated
        '409':
          description: Base or used conversion is locked
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
    delete:
      summary: Delete an unused non-base unit
      responses:
        '204':
          description: Deleted
        '409':
          description: Base or referenced unit
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Unit:
      type: object
      required:
      - id
      - key
      - name
      - symbol
      - scaleToBase
      - offsetToBase
      - isBase
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        symbol:
          type: string
        scaleToBase:
          type: number
          exclusiveMinimum: 0
        offsetToBase:
          type: number
        aliases:
          type: array
          items:
            type: string
        isBase:
          type: boolean
        eventCount:
          type: integer
        defaultEventTypeCount:
          type: integer
    BaseUnitCreate:
      type: object
      required:
      - key
      - name
      - symbol
      properties:
        key:
          type: string
        name:
          type: string
        symbol:
          type: string
        aliases:
          type: array
          items:
            type: string
    UnitTypeCreate:
      type: object
      required:
      - key
      - name
      - baseUnit
      properties:
        key:
          type: string
        name:
          type: string
        description:
          type:
          - string
          - 'null'
        baseUnit:
          $ref: '#/components/schemas/BaseUnitCreate'
    UnitCreate:
      type: object
      required:
      - key
      - name
      - symbol
      - scaleToBase
      properties:
        key:
          type: string
        name:
          type: string
        symbol:
          type: string
        scaleToBase:
          type: number
          exclusiveMinimum: 0
        offsetToBase:
          type: number
          default: 0
        aliases:
          type: array
          items:
            type: string
    UnitUpdate:
      type: object
      properties:
        name:
          type: string
        symbol:
          type: string
        scaleToBase:
          type: number
          exclusiveMinimum: 0
        offsetToBase:
          type: number
        aliases:
          type: array
          items:
            type: string
    UnitType:
      type: object
      required:
      - id
      - key
      - name
      - units
      - eventTypeCount
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        description:
          type:
          - string
          - 'null'
        baseUnit:
          oneOf:
          - $ref: '#/components/schemas/Unit'
          - type: 'null'
        units:
          type: array
          items:
            $ref: '#/components/schemas/Unit'
        eventTypeCount:
          type: integer
...
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      description: Optional unique key for safely retrying POST, PATCH, and DELETE requests. Reusing a key with
        the same request returns the stored response; reusing it with a different request returns 409.
      schema:
        type: string
        minLength: 1
        maxLength: 200